Serialization fix

+++ NAMSC_lib/src/Novel/Action/Audio/ActionAudioSetSounds.cpp
@@ -52,10 +52,9 @@ ActionAudioSetSounds::ActionAudioSetSounds(Event* const parentEvent, const Audio
 void ActionAudioSetSounds::serializableLoad(QDataStream& dataStream)
 {
 	ActionAudio::serializableLoad(dataStream);
-
-	uint soundsSize;
+	size_t soundsSize;
 	dataStream >> soundsSize;
-	for (uint i = 0u; i != soundsSize; ++i)
+	for (size_t i = 0u; i != soundsSize; ++i)
 	{
 		Sound sound;
 		dataStream >> sound;
@@ -67,8 +66,7 @@ void ActionAudioSetSounds::serializableLoad(QDataStream& dataStream)
 void ActionAudioSetSounds::serializableSave(QDataStream& dataStream) const
 {
 	ActionAudio::serializableSave(dataStream);
-
-	dataStream << static_cast<uint>(sounds.size());
+	dataStream << sounds.size();
 	for (const std::pair<const QString, Sound>& sound : sounds)
 		dataStream << sound.second;
 }
